refactor(define-commons): split last services
